Can now insert an object with a [Key] attribute without needing a default value on the DB column.
Added an optional parameter for specifying that the value provided with the [Key] Parameter should be used when performing the insert.  This allows primary key indication (for updates) on tables with no default value for the primary key.

Added a test for the feature added above.  It only runs for Postgres, as SqlCe gets angry about no IDENTITY on Id.

Added the ability to run the tests against different databases for Dapper.Contrib, and implemented running against a Postgres database.  As there is no 'in memory' Postgres feature, so the user will have to set up their own copy.  Due to that, it's commented out by default.