Doesn't redudantly cache for multitenant dbs.
This makes it so the user has the option to pass in regexes that
represent multitent dbs so that dapper can recognize multitenant dbs.
This way a query is not chached separately for each db, rather only
cached once to be shared across the multitenant dbs.
This is done by considering db's with a dbname that match a regex the
same with db's that are equal in every other way accept the dbname, in
the Identity Equals method.
Also connection string is ignored in hashcode so that a set of
multitenant db;s have the same hashcode.