Added support for ICustomQueryParameters to be members of a dictionary that can then be passed into a new DynamicParameters object.
Generally, working with an IDictionary<string,object> and DynamicParameters has worked well.  However, when one of the entries in the dictionary is an ICustomQueryParameter, it would blow up in the SqlMapper.LookupDbType method.  This allows parameters that are part of a dictionary to simply call 'AddParameter' passing in the command and name.