Remove DynamicParamaters.Output() dependency on templates
The current implementation of DynamicParameters.Output(), only updates the output target if the DynamicParameters instance is constructed with a template: i.e. public DynamicParameters(object template);

See: http://stackoverflow.com/questions/39904049/dynamicparameters-outputt-doesnt-map-the-output-to-the-model

As the target is included in the 'Output' definition, I see no reason for this limitation, and propose a simple fix to the DynamicParameters class that removes this dependency.