Revert breaking change, move output callback hooks
As discussed with Nick Craver; the previous changes to the 'callback
hooks' could be a breaking change.  Instead, it should be sufficient to
move when these are hooked-up until after all paramaters have been
added.  That way param.AttachedParam will be non-null making the
'breaking changes' redundant - so they have been reverted.